core(plugin): Rename placeholder to smartcode
