--- src/placeholder/placeholder.js
@@ -1,10 +1,0 @@
-import Plugin from '@ckeditor/ckeditor5-core/src/plugin';
-
-import PlaceholderEditing from './placeholderediting';
-import PlaceholderUI from './placeholderui';
-
-export default class Smartcode extends Plugin {
-	static get requires() {
-		return [ PlaceholderEditing, PlaceholderUI ];
-	}
-}
